Initial version of adding googletest-googlemock to my project

+++ InternetOfThingNetwork/InternetOfThingNetwork/solutionShared/include/GlobalHeader.h
@@ -1,8 +1,8 @@
 
 // Type definitions
 typedef unsigned int IdType;  // type for ID of each node unsigned int/short
-typedef float CoordinateType; // either "float" ot "double"
-typedef float TimeType;		  // time for time 
+typedef double CoordinateType; // either "float" ot "double"
+typedef double TimeType;		  // time for time 
 
 // global constant definitiotns
 #define NODE_MAX_POSSIBLE_SPEED_M_PER_SECOND               350.0*5.0/18.0
